updated generate.py to work with existing notebooks

+++ generate.py
@@ -73,7 +73,24 @@ if __name__ == '__main__':
   )
   args = parser.parse_args()
   for name in args.notebook_names:
-    name = name + ".ipynb"
-    with open(name, 'w') as fp:
-      json.dump(notebook, fp, indent=2)
-      print(f'Generated {os.path.join(cwd, name)}')
+    if name[-6:] != ".ipynb":
+      name = name + ".ipynb"
+
+    if os.path.isfile(name):
+      with open(name) as fp:
+        content = json.load(fp)
+
+      for key, value in content.items():
+        if key != 'cells':
+          content[key] = notebook[key]
+        else:
+          content[key] = notebook[key] + content[key]
+
+      with open(name, mode='w') as f:
+        f.write(json.dumps(content, indent=2))
+        print(f'Added frontmatter to {os.path.join(cwd, name)}')
+
+    else:
+      with open(name, 'w') as fp:
+        json.dump(notebook, fp, indent=2)
+        print(f'Generated {os.path.join(cwd, name)}')
